Batch Parse/Check Script
Added the batch parsing and EAC-CPF checking script.  This script reads
through a list of files on standard input and then parses them into SNAC
Identity Constellation structures in PHP.  Any tags that are extant in
the EAC-CPF, but unparsable by the EACCPFParser, will be displayed on
standard output by this script.  This script may be the baseline for a batch
processing job.