Constellation to/from JSON and Associative Array
Updated all the data classes to allow for conversion back and forth
between PHP and JSON or Associative Arrays.  Every class now extends an
AbstractData class that provides toJSON and fromJSON methods as well as
a default constructor that takes an associative array and fills the data
structure with that array's data.  The classes must then implement a
toArray and fromArray method that appropriately converts the structure
to/from arrays.