Major changes to Data Classes
Added many new data classes, including bioghist, language, legalstatus,
sameas, source, and term.  Each of these is then reflected in the
Constellation.  Updated the parser to correctly build the new format for
these objects.

Also cleaned up much of the PHPDoc documentation to fix errors in the
parameter or return values, or to match style with the rest of the
documentation.