Validation Engine and Operation Validation
Added the validation engine and the operation validator.  The validation
engine loops over the constellation and calls each validator on the
high-level elements of the constellation.  The validator then internally
checks the structures it needs to and returns a list of errors to the
validation engine.  The engine will then throw a SNACValidationException
listing all errors.

Tests for Operation Validator are included to test that a delete or
update may not follow an insert, an insert or update may not follow a
delete, and that anything is possible on update.