Fixed EAC Parser Tests
Updated the EAC-CPF parser to do a just-in-time instantiation of the
default vocabulary, but it now allows for the replacement with a
different vocabulary.  Added a TestVocabulary to the EAC Parser Tests
that don't require the database to allow the parser to be tested without
database interaction.  Fixed the test that checked now invalid portions
of the Constellation object.