Date List fix and Maintenance Status
For some reason, the foreach over dateList in abstract data is failing
if the list is empty.  That is, if the list is empty, it appears to be
null.  I can't figure it out, so I wrapped it in a test for null.

Also, maintenance status is supposed to be a controlled vocabulary term,
so it is now treated as such.  This fixes the to/from JSON bugs we have
been seeing.