Post-Constellation Mapping Bug Fix
Fixed the problem with creating and reconciling objects between the
WebUI-generated Constellations and the server-returned (written)
Constellations.  The PostMapper also now parses down the entire
hierarchy of the Constellation.

Added the logging mechanism to AbstractData, so that any data object
may log information about itself to the appropriate logging channel.