User Upgrades
Added the initial body of the User class, with methods for serializing
to and from array and json.  This initial class has methods that match
data available from Google's OAuth2 API. The WebUI now sends user
information to the server and reads it back from the server.

Cleaned up the web interface relating to users.  Added a profile page
with the base information we get from Google.  We should alter this more
to add groups/roles and permissions from our database.