ID Validator bug fix
The ID validator had problems when it was trying to validate a Source
under a SCM.  Its former implementation didn't allow multiple objects
with the same ID, but that's explicitly happening with Source objects.
So, now it checks duplicates but it handles Source objects under SCM
objects differently.  (It now only checks that those Source objects have
legal IDs from the Constellation)