SCM and Source Fixes
The Sources now have Display Names that are passed all around in the Web
UI.  Also, the EAC Parser now appropriately adds a display name to each
Source.  The Display Names are given to the user via ajax when they are
searching for a source to link to an SCM object.

Now, the WebUI when builds the Constellation, it includes a copy of the
Source object from the Constellation's list as the Citation under an
SCM.  (So, Sources are copied there when they are needed).