Refresh Token on OAuth2
Now, we're asking Google for offline access so that we get a refresh
token from Google.  This allows us to refresh the session (from the
WebUI) when the original session expires.  From the docs, it looks like
we get 6 refreshes using the refresh token before we need to force the
user to log back in.