Moved Authentication to Session Start
Now, authentication only happens when the session begins.  This is a
step back on security, but it allows us to ignore the session expiration
while the user is working.  Without this, the user would be logged out
in 1 hour with no way for us to update their token / session.
Eventually, we'll want to have an additional check that the session is
valid on every call, but we don't want to check Google after the first
session start call.