Edit/Delete Buttons
The edit and delete buttons for non-scm data now function
interdependently.  If a piece is being edited, it cannot be deleted
without undoing the edit.  Similarly, if being deleted, it cannot be
edited without undoing the delete.

Also cleaned up the CSS and set the SCM popups to not change color if
the parent piece is being edited.