Contributors and Select Placeholders
The web ui now supports Contributors attached to NameEntry objects.  By
default, they are "update" when the Name Entry is edited, unless they
are specified as "insert" or "delete" by the user.  Added all plumbing
to add new, remove, and undo, as well as creating objects in the PHP
Constellation and sending/receiving to the Server.

Also added placeholders to the selects across the interface so that
select values can be deselected.