Serializer Fix and Update
Fixed some remaining issues with the seralizer: removed dependence on
global variables, removed dependence on DBUtil instance, added full
maintenance history to the EAC-CPF output.

Moved the serialization and download file creation to the server (from
the web ui) and now the server sends a base64 encoded version of the
file in a JSON object to the web ui for sending to the user.

Updated DBUtil to have an optional flag to read_constellation and its
variants to have DBUtil fill in the maintenance history and maintenance
status out of the version history.