Fixing caching on scripts and css
Now, in debug mode, the display system will append the git commit id as
a GET parameter onto all javascript and css includes from the SNAC
server.  This will force the browser to use the updated versions of the
scripts, but they may be cached for each commit id.

The welcome script no longer needs to load the scripts.  In fact, all it
does now is clear the session.