Large Resource Factoring
Resource data is now stored in a resource cache.  The resource cache has
its own versioning and id system that is taken into account.  Reading
and writing resource relations pulls data from the resource cache.  When
a resource is written, it's inserted and given a new id and version.
They must be written outside of writing a constellation, since writing a
resource relation only writes the ID from a resource.

The view page has been updated to include the resource as a field of the
resource relation data object.  The detailed view and edit pages still
need to be updated.