Deep Merge/Split Lookup Update
This commit contains a deep change in the way the lookup table used.
Before, each id would only map to one "new" id that may be the merged
copy of a constellation.  Now, it stores an implied DAG from id to all
possible ids that it may have been merged/split into.  It does not store
the entire DAG, but only id to final id(s).  That change was propagated
throughout the entire system when the server reads a constellation.