Compare Versions
Updated the view page template to show differences between a new and old
constellation (if they are provided by input) alongside the normal
content of the Constellation.  This allows the view page to show new
changes in green and replacements in red.

Updated the history page to provide links to the view changes version of
the HRT view page.