Save and Send To User
Modified the webui to have a save and send Constellation method that
will use the server's reassign constellation command to send that
constellation to a different user.  The sending user will need the
Change Locks permission to perform this action.

Also, reassigning the constellation now adds a "change locks" status to
the version history to keep the reassign/reviewer note for display to
the editor when the constellation is reassigned.  This is currently
available in the webui on the edit page (entering and reading the
previous note).  The edit page has "Send to Editor" as a way for the
reviewer to send the constellation back to the editor for continued
edits.