Static file serving
The WebUI is now able to serve static files under templating.  It will
read the HTML from a directory, strip out the <title> and <body> and
wrap them into a SNAC template.  It strips out any Javascript and CSS
(includes) from the file and serves up only the HTML content.