Holding Repositories in Display
Added a holding repository relation HIRELATION to Neo4J.  Reading a
Constellation in certain circumstances (display, search, browse) now
checks to see if the identity constellation has a HIRELATION pointing at
it in Neo4J.  If so, a new flag (also new!) is added to the
Constellation denoting it as a Holding Repository.

Creating a new Resource (and updating, hopefully) will update Neo4J to
add the HIRELATION from that resource to its holding repository.  We
still NEED to look at Neo4J on merging Constellations.