Updated edit page to load Constellation by ajax
Added a call to load the entire constellation into memory for reading
and filling out the edit page locally rather than needing to query the
server and hit it for opening each tab.  This makes the page pause a
little when originally opening the edit page, but afterwards everything
else is snappy.  Currently ONLY works for resource relations and DOES
NOT include SCMs.