Improves the Send for Review form.
Modified the send for review form to use a twig tempalte to avoid code
repetition.

This implied changes on the JavaScript in order to use the same field
IDs in all the places that the form was used before.

Modified all the templates that had the duplicated form for the send for
review to use the new twig template.