Adds update SameAs interface to use a control list for URIs
We add a control list to the `vocabulary` table under the type
`external_related_resource_domains`. The interface now is broken up in
two parts a URI base and the element identifier. This change affects
only the frontend the data models for `sameAs` stays the same.

The main idea is to use the control list to prevent common mistakes when
adding data, and to standardise the resources.

The changes in this commit include changes in the `vocabulary` API. The
API used to return an JSON representation of the form:

```JSON
{
results: [
{
id: "400601",
value: "http://datos.bne.es/persona/",
text: "http://datos.bne.es/persona/"
},
{
id: "400595",
value: "http://d-nb.info/gnd/",
text: "http://d-nb.info/gnd/"
},
]
}```

Where the value in the `text` filed was the same as the value for the
vocabulary entry. We want it to use the `description` of the vocabulary
entry as the `text`.

```JSON
[
{
id: "400601",
value: "http://datos.bne.es/persona/",
text: "National Library of Spain"
},
{
id: "400595",
value: "http://d-nb.info/gnd/",
text: "German National Library"
},
{
id: "400598",
value: "http://id.loc.gov/authorities/names/",
text: "Library of Congress/NACO"
},
]```

To accomplish this now the API endpoint now supports a new optional
parameter `use_description` that expects a value `true` or `false`. If
the value exists and it is `true` the returned object will include the
`description`, else it'll return as before. A new call will looks like
the following:

```
http://SNAC_URL/vocabulary?type=external_related_resource_domains&use_description=true
```

Also the changes include the options for the control list to be added in
the SNAC seed data:

```
install/sql_files/vocabulary.sql
```

Updates on the SameAs frontend edit forms

This commit updates how the Same As External Resource form behaves.
The form now displays a control list defined in the vocabulary table.

Because of the control list validation, the users won't be able to
modify URIs that are not in the control list. The URIs that currently
exist on the database that don't comply with the list will be kept in
the database as they are. URIs that don't exist in the control list
should be updated to a URI in the control list.

Currently, the database stores the full URI, and the frontend handles
the split of the baseURI and the URI identifier.