Adds minification and transpilation for specific JavaScripts
Using Webpack module bunlder we are now able to minify the JavaScript
files. We would need to change the structure of our JavaScript source
files in the future but at the moment we'll generate the new minified
JavaScript and add it to `virtualhosts/www/javascript/`.

The current setup, doesn't require any change on our JavaScript only
changes in the directory structure.