Adds same as changes and support for domain description via regex
Fixed bug in Firefox, it didn't support the use of named groups. So
the code was updated to avoid that issue. We are now getting the
matches directly from the array of matches.

Implemented changes requested to the Same As user interface. Details
to the changes requested can be found on issue 256.

Now we support regular expression when defining a Same As
domain. Due to the extraction and matching of the regular expression the
code had to be modified to use Promises in order to be able to wait for
the Domain options obtained via an API. After the select has all the
options we are able to parse the current value to extract the ID using
the regular expression that defines the domain.