Update for ElasticSearch 7
Elastic Search no longer allows a second type for an index, so we need
to split out the all names index from the normal search index.  This
commit separates them and brings the version of the elastic search
Composer library up to the current version.  However, it will require a
new configuration parameter and a refresh of the Elastic Search indices.